+++ psycopg/psycopg/_typeinfo.py
@@ -6,7 +6,7 @@ information to the adapters if needed.
 """
 
 # Copyright (C) 2020 The Psycopg Team
-
+from enum import Enum
 from typing import Any, Dict, Iterator, Optional, overload
 from typing import Sequence, Tuple, Type, TypeVar, Union, TYPE_CHECKING
 
@@ -282,6 +282,42 @@ LEFT JOIN (
     GROUP BY attrelid
 ) a ON a.attrelid = t.typrelid
 WHERE t.oid = %(name)s::regtype
+"""
+
+
+class EnumInfo(TypeInfo):
+    """Manage information about an enum type."""
+
+    __module__ = "psycopg.types.enum"
+
+    def __init__(
+        self,
+        name: str,
+        oid: int,
+        array_oid: int,
+        labels: Sequence[str],
+    ):
+        super().__init__(name, oid, array_oid)
+        self.labels = labels
+        # Will be set by register_enum()
+        self.enum: Optional[Type[Enum]] = None
+
+    @classmethod
+    def _get_info_query(
+        cls, conn: "Union[Connection[Any], AsyncConnection[Any]]"
+    ) -> str:
+        return """\
+SELECT
+    t.typname AS name, t.oid AS oid, t.typarray AS array_oid,
+    array_agg(x.enumlabel) AS labels
+FROM pg_type t
+LEFT JOIN (
+    SELECT e.enumtypid, e.enumlabel
+    FROM pg_enum e
+    ORDER BY e.enumsortorder
+) x ON x.enumtypid = t.oid
+WHERE t.oid = %(name)s::regtype
+GROUP BY t.typname, t.oid, t.typarray
 """